refacotring the code, removing unnecessary buisness logic in the controllers

+++ src/main/java/com/musicUpload/dataHandler/services/SongService.java
@@ -1,10 +1,13 @@
 package com.musicUpload.dataHandler.services;
 
+import com.musicUpload.dataHandler.DTOs.SongDTO;
+import com.musicUpload.dataHandler.details.CustomUserDetails;
 import com.musicUpload.dataHandler.models.ProtectionType;
 import com.musicUpload.dataHandler.models.Song;
 import com.musicUpload.dataHandler.models.User;
 import com.musicUpload.dataHandler.repositories.SongRepository;
 import com.musicUpload.dataHandler.repositories.UserRepository;
+import com.musicUpload.exceptions.UnauthenticatedException;
 import com.musicUpload.util.ImageFactory;
 import com.musicUpload.util.MusicFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,20 +42,25 @@ public class SongService {
         return songRepository.save(song);
     }
 
-    public Song saveSong(String protectionType,
+    public Song saveSong(CustomUserDetails userDetails,
+                         String protectionType,
                          String name,
                          MultipartFile image,
                          MultipartFile songFile,
                          Long userId){
+        if(userDetails == null){
+            throw new UnauthenticatedException();
+        }
+
         User user = userRepository.findById(userId)
                 .orElseThrow(IllegalArgumentException::new);
-        
+
         Song song = new Song();
         song.setUser(user);
         ProtectionType protection = protectionTypeService.getProtectionTypeByName(protectionType)
-                        .orElseThrow(IllegalAccessError::new);
+                .orElseThrow(IllegalAccessError::new);
         song.setProtectionType(protection);
-        
+
         if(name == null){
             throw new IllegalArgumentException();
         }
@@ -74,45 +82,79 @@ public class SongService {
             }
         }
 
-        if(songFile != null && songFile.isEmpty()){
-            try{
-                if(!Objects.requireNonNull(songFile.getContentType()).contains("image")){
+        if(songFile != null && songFile.isEmpty()) {
+            try {
+                if (!Objects.requireNonNull(songFile.getContentType()).contains("image")) {
                     throw new IllegalArgumentException();
                 }
                 String hashedFileName = UUID.randomUUID() + ".mp3";
                 songFile.transferTo(new File(musicFactory.getDirName() + "//" + hashedFileName));
                 musicFactory.deleteFile(song.getNameHashed());
                 song.setNameHashed(hashedFileName);
-            }
-            catch (IOException ioException){
+            } catch (IOException ioException) {
                 //TODO: create a custom exception here
                 throw new IllegalArgumentException();
             }
         }
-        return song;
+        return saveSong(song);
     }
 
-    public void deleteSong(Song song){
+    public Song deleteSong(CustomUserDetails userDetails,
+                           Long id){
+        if(userDetails == null){
+            throw new UnauthenticatedException();
+        }
+
+        Song song = userDetails.getSongs().stream().filter(s -> s.getId().equals(id)).findAny()
+                .orElseThrow(UnauthenticatedException::new);
         imageFactory.deleteFile(song.getImage());
         songRepository.delete(song);
+        return song;
     }
 
     public Optional<Song> findById(Long id){
         return songRepository.findById(id);
     }
 
-    public List<Song> getRandomSongs(){
-        return songRepository.findRandomSongs();
+    public SongDTO findById(CustomUserDetails userDetails,
+                            Long id){
+        Optional<Song> songOptional = findById(id);
+        if(songOptional.isPresent() && !songOptional.get().getProtectionType().getName().equals("PRIVATE")){
+            return SongDTO.of(songOptional.get());
+        }
+
+        if(userDetails != null){
+            Optional<Song> songOptionalForUser = userDetails.getSongs().stream().filter(s -> s.getId().equals(id)).findAny();
+
+            if(songOptionalForUser.isPresent()){
+                return SongDTO.of(songOptionalForUser.get());
+            }
+        }
+        throw new UnauthenticatedException();
+    }
+
+    public List<SongDTO> getRandomSongs(){
+        return songRepository.findRandomSongs().stream()
+                .map(SongDTO::new)
+                .toList();
     }
 
     public Optional<Song> findByNameHashed(String name){
         return songRepository.findByNameHashed(name);
     }
 
-    public void updateSong(Song song,
+    public void updateSong(CustomUserDetails userDetails,
+                           Long id,
                            String protectionType,
                            String name,
                            MultipartFile image){
+        if(userDetails == null){
+            throw new UnauthenticatedException();
+        }
+
+        Song song = userDetails.getSongs().stream().filter(s -> s.getId().equals(id)).findAny()
+                .orElseThrow(UnauthenticatedException::new);
+
         if(protectionType != null){
             Optional<ProtectionType> protectionOpt = protectionTypeService.getProtectionTypeByName(protectionType);
             protectionOpt.ifPresent(song::setProtectionType);
@@ -139,5 +181,13 @@ public class SongService {
         }
 
         songRepository.save(song);
+    }
+
+    public List<SongDTO> getSongs(CustomUserDetails userDetails){
+        if(userDetails == null){
+            throw new UnauthenticatedException();
+        }
+
+        return userDetails.getSongs().stream().map(SongDTO::new).toList();
     }
 }
